+++ google-cloud/container-linux/kubernetes/controllers/controllers.tf
@@ -15,9 +15,10 @@ resource "google_dns_record_set" "etcds" {
 }
 
 # Zones in the region
-data "google_compute_zones" "available" {
+# Only the first 2 should be used. Backend services only accept a fixed number
+# of zonal instance groups and there exist regions with as few as 2 zones.
+data "google_compute_zones" "all" {
   region = "${var.region}"
-  status = "UP"
 }
 
 # Controller instances
@@ -25,7 +26,7 @@ resource "google_compute_instance" "controllers" {
   count = "${var.count}"
 
   name         = "${var.cluster_name}-controller-${count.index}"
-  zone         = "${element(data.google_compute_zones.available.names, count.index)}"
+  zone         = "${element(slice(data.google_compute_zones.all.names, 0, 2), count.index)}"
   machine_type = "${var.machine_type}"
 
   metadata {